# services/news-feed-engine/api/openapi.yaml
openapi: 3.1.0
info:
  title: ElevatedIQ News Feed Engine API
  description: |
    AI-powered news aggregation platform API for multi-platform content ingestion,
    analysis, and video summary generation. Supports white-label multi-tenant deployment.
  version: 1.0.0
  contact:
    name: ElevatedIQ Platform Team
    email: platform@elevatediq.ai
  license:
    name: Proprietary
servers:
  - url: https://api.elevatediq.ai/news-feed/v1
    description: Production
  - url: https://staging-api.elevatediq.ai/news-feed/v1
    description: Staging
  - url: http://127.0.0.1:8096/api/v1
    description: Local Development
...
paths:
...
  /creators:
    get:
      tags: [Creators]
      summary: List creators
      operationId: listCreators
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/platform'
        - name: tier
          in: query
          schema:
            type: string
            enum: [platinum, gold, silver, bronze, unverified]
        - name: verified
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Paginated list of creators
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatorListResponse'
  /creators/{id}:
    get:
      tags: [Creators]
      summary: Get creator by ID
      operationId: getCreator
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Creator details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /creators/{id}/content:
    get:
      tags: [Creators]
      summary: Get creator's content
      operationId: getCreatorContent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - name: type
          in: query
          schema:
            type: string
            enum: [article, video, post, podcast, newsletter]
      responses:
        '200':
          description: Creator's content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentListResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    platform:
      name: platform
      in: query
      schema:
        type: string
        enum: [youtube, twitter, reddit, rss, internal]
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total_items:
          type: integer
        total_pages:
          type: integer
        has_more:
          type: boolean
    Content:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
        creator_id:
          type: string
          format: uuid
        platform:
          type: string
          enum: [youtube, twitter, reddit, rss, internal]
        content_type:
          type: string
          enum: [article, video, post, podcast, newsletter]
        title:
          type: string
        description:
          type: string
        original_url:
          type: string
          format: uri
        thumbnail_url:
          type: string
          format: uri
        summary:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        geo_classification:
          type: string
          enum: [local, regional, national, global]
        sentiment:
          type: number
          format: float
          minimum: -1
          maximum: 1
        quality_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
        engagement_score:
          type: number
          format: float
        view_count:
          type: integer
        like_count:
          type: integer
        comment_count:
          type: integer
        share_count:
          type: integer
        published_at:
          type: string
          format: date-time
        ai_analysis:
          type: object
          properties:
            key_points:
              type: array
              items:
                type: string
            entities:
              type: array
              items:
                type: object
            topics:
              type: array
              items:
                type: string
        creator:
          $ref: '#/components/schemas/Creator'
    Creator:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        platform:
          type: string
        avatar_url:
          type: string
          format: uri
        bio:
          type: string
        tier:
          type: string
          enum: [platinum, gold, silver, bronze, unverified]
        verified_at:
          type: string
          format: date-time
        follower_count:
          type: integer
        engagement_rate:
          type: number
          format: float
...
    ContentListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Content'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    CreatorListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Creator'
        pagination:
          $ref: '#/components/schemas/Pagination'
    CreatorResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Creator'
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string

    # AI Schemas
...